adding new option for https url
