+++ src/main/java/org/jenkinsci/plugins/github_branch_source/SSHCheckoutTrait.java
@@ -133,7 +133,7 @@ public class SSHCheckoutTrait extends SCMSourceTrait {
          */
         @Override
         public Class<? extends SCMSource> getSourceClass() {
-            return GitHubSCMSource.class;
+            return GitHubSCMSourceAbstract.class;
         }
 
         /**
