+++ src/main/java/org/jenkinsci/plugins/github_branch_source/TagDiscoveryTrait.java
@@ -110,7 +110,7 @@ public class TagDiscoveryTrait extends SCMSourceTrait {
          */
         @Override
         public Class<? extends SCMSource> getSourceClass() {
-            return GitHubSCMSource.class;
+            return GitHubSCMSourceAbstract.class;
         }
 
     }
