+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/config-detail.jelly
@@ -1,0 +1,15 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:c="/lib/credentials"
+         xmlns:f="/lib/form"
+         xmlns:scm="/jenkins/scm/api/form">
+  <f:entry title="${%Credentials}" field="credentialsId">
+    <c:select checkMethod="post"/>
+  </f:entry>
+  <f:entry title="${%Repository URL}" field="repositoryURL">
+    <f:textbox/>
+  </f:entry>
+  <f:entry title="${%Behaviours}">
+    <scm:traits field="traits"/>
+  </f:entry>
+
+</j:jelly>
