+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/config-detail_en.properties
@@ -1,0 +1,1 @@
+Behaviours=Behaviours
