+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-apiUri.html
@@ -1,0 +1,4 @@
+<div>
+    The server to connect to. The list of servers is configured in the <strong>Manage Jenkins » Configure
+    Jenkins › GitHub Enterprise Servers</strong> screen.
+</div>
