+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-apiUri_zh_CN.html
@@ -1,0 +1,3 @@
+<div>
+    需要连接的服务器地址。已配置的服务列表在<strong>系统管理 » 系统配置 › GitHub Enterprise Servers</strong>区域。
+</div>
