+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-credentialsId.html
@@ -1,0 +1,22 @@
+<div>
+    <p>
+        Credentials used to <strong>scan</strong> branches and pull requests,
+        <strong>check out</strong> sources and <strong>mark</strong> commit statuses.
+    </p>
+    <p>
+        Note that <strong>only "username with password"</strong> credentials are supported.
+        Existing credentials of other kinds will be filtered out. This is because jenkins
+        exercises GitHub API, and this last one does not support other ways of authentication.
+    </p>
+    <p>
+        If none is given, only the public repositories will be scanned, and commit status
+        will not be set on GitHub.
+    </p>
+    <p>
+        If your organization contains private repositories, then you need to specify
+        a credential from an user who have access to those repositories. This is done
+        by creating a "username with password" credential where the password is
+        <a href="https://github.com/settings/tokens">GitHub personal access tokens</a>.
+        The necessary scope is "repo".
+    </p>
+</div>
