+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-credentialsId_zh_CN.html
@@ -1,0 +1,17 @@
+<div>
+    <p>
+        用于<strong>扫描</strong>分支和 Pull Request，<strong>获取</strong>源代码，<strong>标记</strong>提交状态。
+    </p>
+    <p>
+        注意： 仅支持<strong>“用户名+密码”</strong>类型的凭据。
+        由于 Jenkins 使用的是 GitHub API， 而它仅支持上述类型的授权方式，所以其他类型的凭据将会被过滤掉。
+    </p>
+    <p>
+        如果凭据为空，则只能访问公有代码仓库，且本地的变更信息不能同步到 GitHub。
+    </p>
+    <p>
+        如果您的组织下包含私有库，您需要指定一个对这些库有访问权限的凭据。
+        可以通过创建一个“用户名密+码类型”的凭据，密码通配置 Token 获取：<a href="https://github.com/settings/tokens">GitHub personal access tokens</a>。
+        必要的权限范围为 "repo"。
+    </p>
+</div>
