+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-repoOwner.html
@@ -1,0 +1,3 @@
+<div>
+    <p>Specify the name of the <strong>GitHub Organization</strong> or <strong>GitHub User Account</strong>.</p>
+</div>
