+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-repoOwner_zh_CN.html
@@ -1,0 +1,3 @@
+<div>
+    <p>指定<strong>Github 组织名</strong>或<strong>GitHub 用户名</strong>。</p>
+</div>
