+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-repository.html
@@ -1,0 +1,3 @@
+<div>
+    The repository to scan.
+</div>
