+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-repository_zh_CN.html
@@ -1,0 +1,3 @@
+<div>
+    需要扫描的代码仓库。
+</div>
