+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMSourceHttpsUrl/help-traits.html
@@ -1,0 +1,12 @@
+<div>
+    The behaviours control what is discovered from the GitHub repository. The behaviours are grouped into a number
+    of categories:
+    <dl>
+        <dt>Within repository</dt>
+        <dd>These behaviours determine what gets discovered. <strong>If you do not configure at least one discovery
+            behaviour then nothing will be found!</strong></dd>
+        <dt>General</dt>
+        <dd>These behaviours affect the configuration of each discovered branch / pull request.
+        </dd>
+    </dl>
+</div>
